backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
greenlet==3.2.2
aiosqlite==0.21.0
pydantic-settings==2.9.1
pytest==8.3.5
pytest-asyncio==0.26.0
httpx==0.28.1
